requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
apify-client==1.7.1
python-dotenv==1.0.0
